+++ Demo/SwiftUI/Demo/App/AppDelegate.swift
@@ -4,7 +4,10 @@ class AppDelegate: NSObject, UIApplicationDelegate {
         
     static var orientationLock = UIInterfaceOrientationMask.all
 
-    func application(_ application: UIApplication, supportedInterfaceOrientationsFor window: UIWindow?) -> UIInterfaceOrientationMask {
+    func application(
+        _ application: UIApplication,
+        supportedInterfaceOrientationsFor window: UIWindow?
+    ) -> UIInterfaceOrientationMask {
         return Self.orientationLock
     }
 }