+++ Demo/SwiftUI/Demo/Extensions/Collection+Safe.swift
@@ -1,0 +1,23 @@
+import Foundation
+
+extension Collection {
+
+    subscript(safe index: Index) -> Element? {
+        return indices.contains(index) ? self[index] : nil
+    }
+}
+
+extension MutableCollection {
+    
+    subscript(safe index: Index) -> Element? {
+        get {
+            return indices.contains(index) ? self[index] : nil
+        }
+
+        set(newValue) {
+            if let newValue = newValue, indices.contains(index) {
+                self[index] = newValue
+            }
+        }
+    }
+}